dummy_rproc: don't bail out on missing AP confirmation kick
Currently the AP boots (somehow), however the "callback" IPI doesn't seem
to work - and we get into a situation where the BSP thinks that AP didn't
boot correctly, frees its memory and starts using it - and, thus, quickly
creates a land with unicorns and mermaids.

So just print a warning for now, and rely only on trampoline boot
confirmation.

Signed-off-by: Veaceslav Falico <vfalico@gmail.com>